requirements.txt
telethon>=1.30.3
google-genai==0.9.1
python-dotenv==1.0.0
